pyproject.toml
[project]
name = "hackathon-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.20.0",
    "fastapi>=0.126.0",
    "openai>=2.14.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "uvicorn>=0.38.0",
    "beautifulsoup4>=4.12.3",
]
